+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+* Remove DeaStore provider since the site is no longer online
+
 ## 0.7.2 / 2018-03-18
 
 * Update dependencies